+++ api/admin.py
@@ -1,6 +1,9 @@
 from django.contrib import admin
-from .models import Contact, Country, GeoEntity
+from .models import *
 # Register your models here.
 admin.site.register(Contact.Contact)
 admin.site.register(Country.Country)
 admin.site.register(GeoEntity.GeoEntity)
+admin.site.register(Tree.Tree)
+admin.site.register(Tree.TreeSpecie)
+admin.site.register(Tree.Share)
