--- api/models/TreeSpecie.py
@@ -1,5 +1,0 @@
-from django.db import models
-
-class TreeSpecie(models.Model):
-    commonname = models.CharField(max_length=100, blank=False)
-    sciname = models.CharField(max_length=100, blank=False)
